+++ src/hhnoqual.f
@@ -93,10 +93,10 @@
       use parm
 
       integer :: jrch, ii
-      real :: wtrin, chlin, algin, orgnin, ammoin, nitratin, nitritin
-      real :: orgpin, dispin, cbodin, disoxin, wtmp
-      real :: orgpcon, solpcon, cbodcon, o2con, wtrtot
-      real :: algcon, orgncon, nh3con, no2con, no3con
+      real*8 :: wtrin, chlin, algin, orgnin, ammoin, nitratin, nitritin
+      real*8 :: orgpin, dispin, cbodin, disoxin, wtmp
+      real*8 :: orgpcon, solpcon, cbodcon, o2con, wtrtot
+      real*8 :: algcon, orgncon, nh3con, no2con, no3con
 
       jrch = 0
       jrch = inum1
