+++ src/hruday.f90
@@ -183,15 +183,15 @@
       use parm
 
       integer :: j, sb, ii, iflag
-      real, dimension (mhruo) :: pdvas, pdvs
+      real*8, dimension (mhruo) :: pdvas, pdvs
       character (len=4) :: cropname
 
       !!by zhang print out soil water
       !!===============================    
       integer :: ly
-      real :: sumwater, sumwfsc, sumdepth, sat, wc, dp
-      real :: ssoilwater(100), swfsc(100)
-      real :: soilwater(11), wfsc(11), sum_depth(11) !10, 100, 200, 300, 400, ..., 1000 mm
+      real*8 :: sumwater, sumwfsc, sumdepth, sat, wc, dp
+      real*8 :: ssoilwater(100), swfsc(100)
+      real*8 :: soilwater(11), wfsc(11), sum_depth(11) !10, 100, 200, 300, 400, ..., 1000 mm
       !!by zhang print out soil water
       !!===============================
