+++ src/nuts.f
@@ -24,8 +24,8 @@
 
 !!    ~ ~ ~ ~ ~ ~ END SPECIFICATIONS ~ ~ ~ ~ ~ ~
 
-      real, intent (in) :: u1, u2
-      real, intent (out) :: uu
+      real*8, intent (in) :: u1, u2
+      real*8, intent (out) :: uu
 
       uu = 0.
