+++ src/readres.f
@@ -116,7 +116,7 @@
       character (len=80) :: titldum
       character (len=13) :: resdayo, resmono
       integer :: eof, mon, j
-      real :: resdif, targ, lnvol, res_d50
+      real*8 :: resdif, targ, lnvol, res_d50
 
 !!    initialize local variables
       resdayo = ""
