+++ src/surq_daycn.f
@@ -48,7 +48,7 @@
       use parm
 
       integer :: j
-      real :: r2, bb, pb, cnimp, surfqimp
+      real*8 :: r2, bb, pb, cnimp, surfqimp
 
       j = 0
       j = ihru
